menambahkan top application

+++ backend/src/main/java/com/traccapp/demo/payload/response/TopApplicationsResponse.java
@@ -7,5 +7,5 @@ import lombok.Data;
 @AllArgsConstructor
 public class TopApplicationsResponse {
     private String application;
-    private int count;
+    private long count;
 }
